Consolidate the various PbfLcMsRun spectrum data readers that looked for spectrum data into one - Removed separate reading function for GetMs1Spectrum, ReadSpectrum(offset, ...) now includes a parameter to try to detect when it should return an Ms1Spectrum - Removed separate reading function for Get IsolationWindow, it now uses ReadSpectrum(offset, includePeaks=false, ...) to read the spectrum, and returns the included IsolationWindow - ReadSpectrum(BinaryReader, ...) is now the sole function that reads and returns Spectrum and derived objects (including contained objects) - GetPeakMetaDataForSpectrum(...) also reads a Spectrum object, but what is returns is completely non-compatible with Spectrum or any derived object. It also is only used during Pbf creation.

+++ InformedProteomics.Backend/MassSpecData/InMemoryLcMsRun.cs
@@ -359,7 +359,7 @@ namespace InformedProteomics.Backend.MassSpecData
             var ms1ScanNums = GetMs1ScanVector();
             var ms1ScanIndex = Array.BinarySearch(ms1ScanNums, scanNum);
             if (ms1ScanIndex < 0) return null;
-            return new Ms1Spectrum(scanNum, (ushort)ms1ScanIndex, spec.Peaks);
+            return new Ms1Spectrum(scanNum, ms1ScanIndex, spec.Peaks);
         }
 
         public override IsolationWindow GetIsolationWindow(int scanNum)
